Dev version 0.8.7

+++ src/app/api/applicationDelete/route.ts
@@ -6,12 +6,16 @@ export async function DELETE(req: Request) {
   const post_id = url.searchParams.get('post_id')
   const userId = url.searchParams.get('user_id')
 
+  if (!post_id && !userId) {
+    return new Response(JSON.stringify({ message: '잘못된 요청입니다.' }), { status: 404 })
+  }
+
   try {
     const res =
       await sql`DELETE FROM applicants_list WHERE user_id = ${userId} AND post_id = ${post_id}`
 
     return new Response(JSON.stringify({ message: '성공' }), {
-      status: 201,
+      status: 200,
     })
   } catch (error) {
     console.error(error)
